backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
openpyxl
reportlab
faker
python-dotenv
